Split the read-receipt tests into logical units
